+++ packages/react-mui-audit-log-viewer/src/components/AuditPreviewRow/index.tsx
@@ -24,6 +24,19 @@ const AuditPreviewRow: FC<Props> = ({
   // @ts-ignore - An id is needed for mui data grid so one is added.
   const rowId: string = record.id;
 
+  let date;
+  if (record?.timestamp) {
+    date = new Date(record?.timestamp);
+    date = date.toLocaleDateString(undefined, {
+      year: "numeric",
+      month: "2-digit",
+      day: "2-digit",
+      hour: "2-digit",
+      minute: "2-digit",
+      second: undefined,
+    });
+  }
+
   return (
     <Stack
       padding={1}
@@ -83,6 +96,13 @@ const AuditPreviewRow: FC<Props> = ({
               <StringField
                 title="Source"
                 value={record.source}
+                uniqueId={rowId}
+              />
+            )}
+            {!visibilityModel?.timestamp && (
+              <StringField
+                title="Timestamp"
+                value={date ?? "-"}
                 uniqueId={rowId}
               />
             )}